+++ src/icv_taxonomy/migrations/0003_alter_term_vocabulary_alter_termassociation_term_and_more.py
@@ -8,6 +8,8 @@ from django.db import migrations, models
 class Migration(migrations.Migration):
     dependencies = [
         ("icv_taxonomy", "0002_term_scoped_path"),
+        migrations.swappable_dependency(getattr(settings, "ICV_TAXONOMY_VOCABULARY_MODEL", "icv_taxonomy.Vocabulary")),
+        migrations.swappable_dependency(getattr(settings, "ICV_TAXONOMY_TERM_MODEL", "icv_taxonomy.Term")),
     ]
 
     operations = [